refactor: 循環import解消、aggregatorとAggregationContextの責務を明確化
SQLヘルパー・前処理(computeTotalN/fetchCrossHeaders)をAggregationContext側に移動し、
aggregator.tsを型定義+薄いエントリポイントに整理。static create()ファクトリを導入。
