+++ src/lib/aggregator.ts
@@ -44,45 +44,12 @@ export interface Query {
   cross_cols: QuestionDef[];
 }
 
-// --- SQL ユーティリティ（AggregationContext からも利用） ---
-
-export function esc(name: string): string {
-  return name.replace(/"/g, '""');
-}
-
-export function weightExpr(weightCol: string): string {
-  return weightCol
-    ? `SUM(TRY_CAST("${esc(weightCol)}" AS DOUBLE))`
-    : `COUNT(*)::DOUBLE`;
-}
-
-/** MA列の重み付きカウント式 */
-export function maWeightedCountExpr(maCol: string, weightCol: string): string {
-  return weightCol
-    ? `SUM(CASE WHEN "${esc(maCol)}" IN ('1','true') THEN TRY_CAST("${esc(weightCol)}" AS DOUBLE) ELSE 0 END)`
-    : `COUNT(CASE WHEN "${esc(maCol)}" IN ('1','true') THEN 1 END)::DOUBLE`;
-}
-
-// --- クロスヘッダーキャッシュ ---
-
-export type CrossHeaderCache = Map<
-  string,
-  { headers: Array<{ label: string; n: number }>; crossValues: string[] }
->;
-
 /** 集計のエントリポイント。conn上のsurveyビューに対して全設問を集計する */
 export async function aggregate(
   conn: duckdb.AsyncDuckDBConnection,
   payload: Query
 ): Promise<AggResult[]> {
-  const totalN = await computeTotalN(conn, payload.weight_col);
-  const crossCols = payload.cross_cols ?? [];
-  const crossHeaderCache =
-    crossCols.length > 0
-      ? await fetchCrossHeaders(conn, crossCols, payload.weight_col)
-      : (new Map() as CrossHeaderCache);
-
-  const ctx = new AggregationContext(conn, payload.weight_col, totalN, crossCols, crossHeaderCache);
+  const ctx = await AggregationContext.create(conn, payload);
 
   const results: AggResult[] = [];
   for (const q of payload.questions) {
@@ -94,62 +61,4 @@ export async function aggregate(
   }
 
   return results;
-}
-
-// --- 内部ユーティリティ ---
-
-async function computeTotalN(
-  conn: duckdb.AsyncDuckDBConnection,
-  weightCol: string
-): Promise<number> {
-  const sql = weightCol
-    ? `SELECT COALESCE(SUM(TRY_CAST("${esc(weightCol)}" AS DOUBLE)), 0) AS n FROM survey`
-    : `SELECT COUNT(*) AS n FROM survey`;
-  const result = await conn.query(sql);
-  return Number(result.toArray()[0].n);
-}
-
-async function fetchCrossHeaders(
-  conn: duckdb.AsyncDuckDBConnection,
-  crossCols: QuestionDef[],
-  weightCol: string
-): Promise<CrossHeaderCache> {
-  const cache: CrossHeaderCache = new Map();
-
-  for (const crossQ of crossCols) {
-    if (crossQ.type === "SA") {
-      const col = crossQ.column;
-      const sql = `
-        SELECT
-          "${esc(col)}" AS cv,
-          ${weightExpr(weightCol)} AS n
-        FROM survey
-        WHERE "${esc(col)}" IS NOT NULL
-          AND "${esc(col)}" != ''
-        GROUP BY "${esc(col)}"
-        ORDER BY
-          TRY_CAST("${esc(col)}" AS DOUBLE) NULLS LAST,
-          "${esc(col)}" ASC
-      `;
-      const result = await conn.query(sql);
-      const headers = result
-        .toArray()
-        .map((r) => ({ label: String(r.cv), n: Number(r.n) }));
-      cache.set(crossQ.column, { headers, crossValues: headers.map((h) => h.label) });
-    } else {
-      // MA軸: 各itemカラムごとに該当者数を算出
-      const selectClauses = crossQ.columns.map((col, i) =>
-        `${maWeightedCountExpr(col, weightCol)} AS c${i}`
-      );
-      const sql = `SELECT ${selectClauses.join(", ")} FROM survey`;
-      const result = await conn.query(sql);
-      const row = result.toArray()[0];
-      const headers = crossQ.columns.map((col, i) => ({
-        label: col,
-        n: Number(row[`c${i}`] ?? 0),
-      }));
-      cache.set(crossQ.prefix, { headers, crossValues: headers.map((h) => h.label) });
-    }
-  }
-  return cache;
 }
